Added the way to handle post request from signup page.
Now signupuser() will handle GET and POST request differently.
User objects are created and saved in the database if the two passwords the user provides matches.
However, if it doesn't, it displays an error message in the terminal.
After submitting the form, the user will be saved in the database but Django will throw error since we don't actually provide a webpage to show after the action taken.