Basic login functionality added.
Created current/ path in urls.py and current_todos() in views.py.
If all goes well in the signup page, the user will be logged in and will be redirected to current/ in urls.py
and this will refer to current_todos() in views.py which will render current_todos.html template.