Display todos of the user in the current todos page.
The current_todos() now passes the todo objects which is made by the logged in user and not yet completed to the template current_todos.html.
If it is marked important then it is shown in bold and if there is some additional info then it is shown.
Also added 'Create Todo!' button at the end of current_todos to go to the Create todo page and 'Cancel' button at the end of the Create todo page to go back to current todos page.