Todos can be viewed and edited on click.
Added hyperlinks to every todo in the current_todos page which upon clicking, opens the view_todo url and passes the primary key of the todo object as an url argument.
The view_todo then takes the argument and passes to the view_todo() which fetches the object from the database if the primary key is valid and if the user of the requested object
is the same as the user currently logged in, so that the user cannot view any other user's views. Then, it creates a form [TodoForm] based on the existing instance's data and displays
it to the view_todo.html template page. If the user edits the information and clicks the save button, it is updated and saved in the database and it redirects the user back to the
current_todos page. Also added buttons in the home page to go to the create_todo and current_todos page.