Todos can be 'completed' through the view page.
In the view_todo page, if the user clicks the completed button, it makes a POST to the complete_todo view and passes the primary key of the todo object along with it.
Then the post is handled by complete_todo(), which sets the completion time of the todo object as the current time and then saves it to the database.
Finally it redirects the user to the current_todos view.
But, the completed todo object is gone. Becuase in the current_todos() it only fetches the todo objects which have the completed attribute as null. So, since in the  complete_todo()
we have set the time to the current time, therefore it is no more null, and hence it is not fetched by current_todos().