Todos can be 'deleted' through the view page.
In the view_todo page, if the user clicks the completed button, it makes a POST to the delete_todo view and passes the primary key of the todo object along with it.
Then the post is handled by delete_todo(), which fetches the object using the given primary key and deletes it from the database.
Finally it redirects the user to the current_todos view.
But, the deleted todo object is gone. Becuase in the current_todos() it only fetches the todo objects which have the valid primary key. So, since in the deleted_todo() we have
deleted the object, the primary key of the object is not available in the database, thus current_todos() couldn't fetch it.o