Completed todos are displyed in most recent order.
When we click the 'completed todos' button on current todos page, it takes the user to completed_todos view and displays all the todo objects that are completed and its
corresponding date & time of completion in most recent order (LIFO).
We can go back and forth between current_todos and completed_todos page.