delete_todo,complete_todo,view_todo,completed_todos,current_todos,create_todos and logoutuser views cannot be accessed by the user if not logged in.
In views.py, imported Django's built in Decorator that does not allows the user to access certain pages unless he/she is logged in. We just put it above the functions for which we want the
users to be logged in order to access those pages.
However, if they try to access those pages without being logged in he/she will be redirected to loginuser page. For this, I went to settings.py and added a variable LOGIN_URL and set it to
'/login'